requirements.txt
# Core packages
pandas==2.2.1
numpy==1.26.4
pyyaml==6.0.1

# Database & API connectors (use only those you need)
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
requests==2.31.0

# AWS interaction
boto3==1.34.76

# Data validation
great_expectations==0.18.13

# ETL orchestration (if using Airflow)
apache-airflow==2.9.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-amazon==8.14.0

# Optional: Scheduling locally
schedule==1.2.1

# For development and testing
pytest==8.1.1
